requirements.txt
flask==3.1.2
openai==2.26.0
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.0.2
playwright==1.55.0
python-dotenv==1.2.2
gunicorn==25.1.0
